requirements.txt
Django==1.9.1
facebook-sdk==0.4.0
psycopg2==2.6.1
requests==2.9.1
wheel==0.26.0
